requirements.txt
absl-py==2.1.0
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astunparse==1.6.3
atomicwrites==1.4.1
attrs==25.1.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cryptography==44.0.2
cssselect2==0.8.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.79.0
filelock==3.18.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.69.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.7.4.2
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
lxml==5.3.2
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==10.6.0
mpmath==1.3.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.28.0
networkx==3.2.1
numpy==1.26.4
openai==1.65.3
opencv-python==4.6.0.66
opt_einsum==3.4.0
optree==0.14.0
oscrypto==1.3.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.4.2
pathlib2==2.3.5
pillow==11.1.0
pluggy==0.13.1
proto-plus==1.26.0
protobuf==5.29.4
py==1.11.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==1.9.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
PyMySQL==1.1.1
pyotp==2.9.0
pyparsing==3.2.1
pypdf==5.4.0
pytest==5.4.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.5
python-slugify==8.0.4
pytz==2020.1
PyYAML==6.0.2
qrcode==8.1
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
runwayml==2.3.3
scikit-learn==1.0.2
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.19.1
streamlit==1.42.2
svglib==1.5.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.4.0
toml==0.10.2
torch==2.6.0+cpu
torchaudio==2.6.0+cpu
torchvision==0.21.0+cpu
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.0
tzdata==2025.1
tzlocal==5.3.1
unicodecsv==0.14.1
uritemplate==4.1.1
uritools==4.0.3
urllib3==1.25.9
uvicorn==0.18.2
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xhtml2pdf==0.2.17
zipp==3.21.0
